requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
